+++ smog/const.py
@@ -1,4 +1,4 @@
-VERSION = "v0.0.3-a"
+VERSION = "v0.0.3"
 
 # DEFAULT_MEDIA_DOC = "~/media-documents"